+++ ci_cd/pipe_stack/main.tf
@@ -5,6 +5,7 @@ module "code_pipe" {
   pipe_source_artifact_format_default          = var.pipe_source_artifact_format_default
   pipe_source_branch_default                   = var.pipe_source_branch_default
   pipe_source_code_star_connection_key_default = var.pipe_source_code_star_connection_key_default
+  pipe_source_detect_changes_default           = var.pipe_source_detect_changes_default
   pipe_source_repository_id_default            = var.pipe_source_repository_id_default
   pipe_stage_category_default                  = var.pipe_stage_category_default
   pipe_webhook_enable_github_hook_default      = var.pipe_webhook_enable_github_hook_default