+++ elb/target_group/locals.tf
@@ -1,6 +1,6 @@
 module "name_map" {
   source                          = "../../name_map"
-  name_include_app_fields_default = var.target_name_include_app_fields_default
+  name_include_app_fields_default = false
   name_infix_default              = var.target_name_infix_default
   name_map                        = local.l0_map
   std_map                         = var.std_map