+++ iam/role/code_build/variables.tf
@@ -28,14 +28,26 @@ variable "code_star_connection_key" {
   description = "If provided, permission to use the connection will be attached"
 }
 
+variable "ecr_data_map" {
+  type = map(object({
+    iam_policy_arn_map = object({
+      read_write = string
+    })
+  }))
+  default     = null
+  description = "Must be provided if image key is specified"
+}
+
 variable "map_policy" {
   type = object({
+    ecr_repo_key                 = optional(string) # The key of the image to build. If specified read/write for the image will be granted.
     role_policy_attach_arn_map   = optional(map(string))
     role_policy_create_json_map  = optional(map(string))
     role_policy_inline_json_map  = optional(map(string))
     role_policy_managed_name_map = optional(map(string))
   })
   default = {
+    ecr_repo_key                 = null
     role_policy_attach_arn_map   = null
     role_policy_create_json_map  = null
     role_policy_inline_json_map  = null