+++ secret/random/main.tf
@@ -1,10 +1,12 @@
 module "this_param" {
   source                                      = "../../ssm/parameter_secret"
+  name_append_default                         = var.name_append_default
+  name_include_app_fields_default             = var.name_include_app_fields_default
+  name_infix_default                          = var.name_infix_default
+  name_prefix_default                         = var.name_prefix_default
+  name_suffix_default                         = var.name_suffix_default
+  name_prepend_default                        = var.name_prepend_default
   param_map                                   = local.create_param_map
-  param_name_include_app_fields_default       = var.secret_name_include_app_fields_default
-  param_name_infix_default                    = var.secret_name_infix_default
-  param_name_prefix_default                   = var.secret_name_prefix_default
-  param_name_suffix_default                   = var.secret_name_suffix_default
   policy_access_list_default                  = var.policy_access_list_default
   policy_create_default                       = var.policy_create_default
   policy_name_append_default                  = var.policy_name_append_default
@@ -22,6 +24,12 @@ module "this_param" {
 
 module "this_secret" {
   source                                      = "../../sm/secret"
+  name_append_default                         = var.name_append_default
+  name_include_app_fields_default             = var.name_include_app_fields_default
+  name_infix_default                          = var.name_infix_default
+  name_prefix_default                         = var.name_prefix_default
+  name_suffix_default                         = var.name_suffix_default
+  name_prepend_default                        = var.name_prepend_default
   policy_access_list_default                  = var.policy_access_list_default
   policy_create_default                       = var.policy_create_default
   policy_name_append_default                  = var.policy_name_append_default
@@ -30,10 +38,6 @@ module "this_secret" {
   policy_name_prepend_default                 = var.policy_name_prepend_default
   policy_name_suffix_default                  = var.policy_name_suffix_default
   secret_map                                  = local.create_secret_map
-  secret_name_include_app_fields_default      = var.secret_name_include_app_fields_default
-  secret_name_infix_default                   = var.secret_name_infix_default
-  secret_name_prefix_default                  = var.secret_name_prefix_default
-  secret_name_suffix_default                  = var.secret_name_suffix_default
   secret_random_init_key_default              = var.secret_random_init_key_default
   secret_random_init_map_default              = var.secret_random_init_map_default
   secret_random_init_type_default             = var.secret_random_init_type_default