+++ sm/secret/locals.tf
@@ -1,10 +1,12 @@
 module "name_map" {
   source                          = "../../name_map"
-  name_include_app_fields_default = var.secret_name_include_app_fields_default
-  name_infix_default              = var.secret_name_infix_default
+  name_append_default             = var.name_append_default
+  name_include_app_fields_default = var.name_include_app_fields_default
+  name_infix_default              = var.name_infix_default
   name_map                        = local.l0_map
-  name_prefix_default             = var.secret_name_prefix_default
-  name_suffix_default             = var.secret_name_suffix_default
+  name_prefix_default             = var.name_prefix_default
+  name_prepend_default            = var.name_prepend_default
+  name_suffix_default             = var.name_suffix_default
   std_map                         = var.std_map
 }