+++ ssm/parameter_secret/main.tf
@@ -1,11 +1,11 @@
 module "initial_value" {
   source                                      = "../../secret/init_value"
-  secret_map                                  = local.param_map
+  secret_map                                  = local.lx_map
   secret_random_special_character_set_default = var.secret_random_special_character_set_default
 }
 
 resource "aws_ssm_parameter" "this_param" {
-  for_each        = local.param_map
+  for_each        = local.lx_map
   allowed_pattern = null
   data_type       = "text"
   insecure_value  = null